More work

+++ lib/lir/x64/asm/emitters/CMovGPEmit.h
@@ -45,11 +45,11 @@ private:
         unimplemented();
     }
 
-    void emit(aasm::GPReg out, aasm::GPReg in1, std::int32_t in2) override {
+    void emit(aasm::GPReg out, aasm::GPReg in1, std::int64_t in2) override {
         unimplemented();
     }
 
-    void emit(aasm::GPReg out, std::int32_t in1, aasm::GPReg in2) override {
+    void emit(aasm::GPReg out, std::int64_t in1, aasm::GPReg in2) override {
         if (out == in2) {
             m_as.copy(m_size, in1, m_temporal_regs.gp_temp1());
             m_as.cmov(m_size, m_cond_type, m_temporal_regs.gp_temp1(), out);
@@ -60,7 +60,7 @@ private:
         m_as.cmov(m_size, aasm::invert(m_cond_type), in2, out);
     }
 
-    void emit(const aasm::GPReg out, const std::int32_t in1, const std::int32_t in2) override {
+    void emit(const aasm::GPReg out, const std::int64_t in1, const std::int64_t in2) override {
         if (in1 == in2) {
             m_as.copy(m_size, in1, out);
             return;
@@ -71,11 +71,11 @@ private:
         m_as.cmov(m_size, m_cond_type, m_temporal_regs.gp_temp1(), out);
     }
 
-    void emit(aasm::GPReg out, std::int32_t in1, const aasm::Address &in2) override {
+    void emit(aasm::GPReg out, std::int64_t in1, const aasm::Address &in2) override {
         unimplemented();
     }
 
-    void emit(aasm::GPReg out, const aasm::Address &in1, std::int32_t in2) override {
+    void emit(aasm::GPReg out, const aasm::Address &in1, std::int64_t in2) override {
         unimplemented();
     }
 
@@ -95,23 +95,23 @@ private:
         unimplemented();
     }
 
-    void emit(const aasm::Address &out, aasm::GPReg in1, std::int32_t in2) override {
+    void emit(const aasm::Address &out, aasm::GPReg in1, std::int64_t in2) override {
         unimplemented();
     }
 
-    void emit(const aasm::Address &out, std::int32_t in1, std::int32_t in2) override {
+    void emit(const aasm::Address &out, std::int64_t in1, std::int64_t in2) override {
         unimplemented();
     }
 
-    void emit(const aasm::Address &out, const aasm::Address &in1, std::int32_t in2) override {
+    void emit(const aasm::Address &out, const aasm::Address &in1, std::int64_t in2) override {
         unimplemented();
     }
 
-    void emit(const aasm::Address &out, std::int32_t in1, const aasm::Address &in2) override {
+    void emit(const aasm::Address &out, std::int64_t in1, const aasm::Address &in2) override {
         unimplemented();
     }
 
-    void emit(const aasm::Address &out, std::int32_t in1, aasm::GPReg in2) override {
+    void emit(const aasm::Address &out, std::int64_t in1, aasm::GPReg in2) override {
         unimplemented();
     }
