+++ lib/lir/x64/asm/emitters/LoadByIdxIntEmit.h
@@ -31,23 +31,23 @@ private:
         unimplemented();
     }
 
-    void emit(const aasm::GPReg out, const aasm::GPReg in1, const std::int32_t in2) override {
+    void emit(const aasm::GPReg out, const aasm::GPReg in1, const std::int64_t in2) override {
         m_as.mov(m_size, aasm::Address(in1, m_size * in2), out);
     }
 
-    void emit(aasm::GPReg out, std::int32_t in1, aasm::GPReg in2) override {
+    void emit(aasm::GPReg out, std::int64_t in1, aasm::GPReg in2) override {
         unimplemented();
     }
 
-    void emit(aasm::GPReg out, std::int32_t in1, std::int32_t in2) override {
+    void emit(aasm::GPReg out, std::int64_t in1, std::int64_t in2) override {
         unimplemented();
     }
 
-    void emit(aasm::GPReg out, std::int32_t in1, const aasm::Address &in2) override {
+    void emit(aasm::GPReg out, std::int64_t in1, const aasm::Address &in2) override {
         unimplemented();
     }
 
-    void emit(aasm::GPReg out, const aasm::Address &in1, std::int32_t in2) override {
+    void emit(aasm::GPReg out, const aasm::Address &in1, std::int64_t in2) override {
         unimplemented();
     }
 
@@ -67,23 +67,23 @@ private:
         unimplemented();
     }
 
-    void emit(const aasm::Address &out, aasm::GPReg in1, std::int32_t in2) override {
+    void emit(const aasm::Address &out, aasm::GPReg in1, std::int64_t in2) override {
         unimplemented();
     }
 
-    void emit(const aasm::Address &out, std::int32_t in1, std::int32_t in2) override {
+    void emit(const aasm::Address &out, std::int64_t in1, std::int64_t in2) override {
         unimplemented();
     }
 
-    void emit(const aasm::Address &out, const aasm::Address &in1, std::int32_t in2) override {
+    void emit(const aasm::Address &out, const aasm::Address &in1, std::int64_t in2) override {
         unimplemented();
     }
 
-    void emit(const aasm::Address &out, std::int32_t in1, const aasm::Address &in2) override {
+    void emit(const aasm::Address &out, std::int64_t in1, const aasm::Address &in2) override {
         unimplemented();
     }
 
-    void emit(const aasm::Address &out, std::int32_t in1, aasm::GPReg in2) override {
+    void emit(const aasm::Address &out, std::int64_t in1, aasm::GPReg in2) override {
         unimplemented();
     }
