+++ lib/lir/x64/asm/visitors/GPBinaryVisitor.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include "lir/x64/lir_frwd.h"
-
 #include "lir/x64/asm/GPOp.h"
 
 class GPBinaryVisitor {
@@ -13,23 +11,23 @@ public:
     virtual void emit(aasm::GPReg out, const aasm::Address& in1, aasm::GPReg in2) = 0;
     virtual void emit(aasm::GPReg out, const aasm::Address& in1, const aasm::Address& in2) = 0;
 
-    virtual void emit(aasm::GPReg out, aasm::GPReg in1, std::int32_t in2) = 0;
-    virtual void emit(aasm::GPReg out, std::int32_t in1, aasm::GPReg in2) = 0;
-    virtual void emit(aasm::GPReg out, std::int32_t in1, std::int32_t in2) = 0;
+    virtual void emit(aasm::GPReg out, aasm::GPReg in1, std::int64_t in2) = 0;
+    virtual void emit(aasm::GPReg out, std::int64_t in1, aasm::GPReg in2) = 0;
+    virtual void emit(aasm::GPReg out, std::int64_t in1, std::int64_t in2) = 0;
 
-    virtual void emit(aasm::GPReg out, std::int32_t in1, const aasm::Address& in2) = 0;
-    virtual void emit(aasm::GPReg out, const aasm::Address& in1, std::int32_t in2) = 0;
+    virtual void emit(aasm::GPReg out, std::int64_t in1, const aasm::Address& in2) = 0;
+    virtual void emit(aasm::GPReg out, const aasm::Address& in1, std::int64_t in2) = 0;
 
     virtual void emit(const aasm::Address& out, aasm::GPReg in1, aasm::GPReg in2) = 0;
     virtual void emit(const aasm::Address& out, aasm::GPReg in1, const aasm::Address& in2) = 0;
     virtual void emit(const aasm::Address& out, const aasm::Address& in1, aasm::GPReg in2) = 0;
     virtual void emit(const aasm::Address& out, const aasm::Address& in1, const aasm::Address& in2) = 0;
 
-    virtual void emit(const aasm::Address& out, aasm::GPReg in1, std::int32_t in2) = 0;
-    virtual void emit(const aasm::Address& out, std::int32_t in1, std::int32_t in2) = 0;
-    virtual void emit(const aasm::Address& out, const aasm::Address& in1, std::int32_t in2) = 0;
-    virtual void emit(const aasm::Address& out, std::int32_t in1, const aasm::Address& in2) = 0;
-    virtual void emit(const aasm::Address& out, std::int32_t in1, aasm::GPReg in2) = 0;
+    virtual void emit(const aasm::Address& out, aasm::GPReg in1, std::int64_t in2) = 0;
+    virtual void emit(const aasm::Address& out, std::int64_t in1, std::int64_t in2) = 0;
+    virtual void emit(const aasm::Address& out, const aasm::Address& in1, std::int64_t in2) = 0;
+    virtual void emit(const aasm::Address& out, std::int64_t in1, const aasm::Address& in2) = 0;
+    virtual void emit(const aasm::Address& out, std::int64_t in1, aasm::GPReg in2) = 0;
 
     template<std::derived_from<GPBinaryVisitor> Vis>
     static void dispatch(Vis &visitor, const GPVReg &out, const GPOp &in1, const GPOp &in2) {
